Microtime addded

+++ app/jobs/chat_message_broadcast_job.rb
@@ -7,7 +7,7 @@ class ChatMessageBroadcastJob < ApplicationJob
       event: 'message',
   		content: message.content, 
   		username: message.user.username, 
-  		created_at: message.created_at.utc.to_i
+  		created_at: message.created_at.utc.to_f
   	}
     ActionCable.server.broadcast "chat", payload
   end
